+++ src/index.ts
@@ -2,7 +2,7 @@ export {
   useExposureForNode,
   useExposureForPage,
   useStayTime,
-} from "./exposureBased";
+} from "./exposureBased/index.js";
 export type {
   IUseExposureForNodeOptions,
   IUseExposureForNodeReturn,
@@ -22,20 +22,22 @@ export type {
   PromiseType,
   StateFromFunctionReturningPromise,
 } from "./useAsyncFn.js";
-export { default as useAsyncFn } from "./useAsyncFn.js";
-export { default as useDebounce } from "./useDebounce.js";
-export { default as useEventListener } from "./useEventListener.js";
-export type { UseInputOptions } from "./useInput.js";
-export { default as useInput } from "./useInput.js";
-export { default as useIntersection } from "./useIntersection.js";
-export { default as useMainThreadImperativeHandle } from "./useMainThreadImperativeHandle.js";
-export { default as useTapAway } from "./useTapAway.js";
-export type { TapLockDirection, UseTapLockOptions } from "./useTapLock.js";
-export { default as useTapLock } from "./useTapLock.js";
-export type { UseTimeoutReturn } from "./useTimeout.js";
-export { default as useTimeout } from "./useTimeout.js";
-export type { UseTimeoutFnReturn } from "./useTimeoutFn.js";
-export { default as useTimeoutFn } from "./useTimeoutFn.js";
+export { default as useAsyncFn } from './useAsyncFn.js';
+export { default as useDebounce } from './useDebounce.js';
+export { default as useEventListener } from './useEventListener.js';
+export type { UseInputOptions } from './useInput.js';
+export { default as useInput } from './useInput.js';
+export { default as useIntersection } from './useIntersection.js';
+export { default as useMainThreadImperativeHandle } from './useMainThreadImperativeHandle.js';
+export type { CustomPointerEvent, CustomPointerEventMT } from './usePointerEvent.js';
+export { default as usePointerEvent } from './usePointerEvent.js';
+export type { TapLockDirection, UseTapLockOptions } from './useTapLock.js';
+export { default as useTapLock } from './useTapLock.js';
+export type { UseTimeoutReturn } from './useTimeout.js';
+export { default as useTimeout } from './useTimeout.js';
+export type { UseTimeoutFnReturn } from './useTimeoutFn.js';
+export { default as useTimeoutFn } from './useTimeoutFn.js';
+export { default as useTouchEmulation, default as useTouchEvent } from './useTouchEmulation.js';
 export type {
   Direction,
   UseVelocityOptions,