+++ src/react-use.ts
@@ -1,33 +1,31 @@
-import {
-  createMemo as _createMemo,
-  useBoolean as _useBoolean,
-  useCounter as _useCounter,
-  useDebounce as _useDebounce,
-  useDefault as _useDefault,
-  useEffectOnce as _useEffectOnce,
-  useError as _useError,
-  useLatest as _useLatest,
-  useLifecycles as _useLifecycles,
-  useMap as _useMap,
-  useMountedState as _useMountedState,
-  useNumber as _useNumber,
-  usePrevious as _usePrevious,
-  useQueue as _useQueue,
-  useSet as _useSet,
-  useSetState as _useSetState,
-  useThrottle as _useThrottle,
-  useThrottleFn as _useThrottleFn,
-  useToggle as _useToggle,
-  useUnmount as _useUnmount,
-  useUnmountPromise as _useUnmountPromise,
-  useUpdateEffect as _useUpdateEffect,
-} from 'react-use';
-import { backgroundOnlyFactory as factory } from './backgroundOnlyFactory';
+import { default as _createMemo } from 'react-use/esm/factory/createMemo.js';
+import { default as _useBoolean } from 'react-use/esm/useBoolean.js';
+import { default as _useCounter } from 'react-use/esm/useCounter.js';
+import { default as _useDebounce } from 'react-use/esm/useDebounce.js';
+import { default as _useDefault } from 'react-use/esm/useDefault.js';
+import { default as _useEffectOnce } from 'react-use/esm/useEffectOnce.js';
+import { default as _useError } from 'react-use/esm/useError.js';
+import { default as _useLatest } from 'react-use/esm/useLatest.js';
+import { default as _useLifecycles } from 'react-use/esm/useLifecycles.js';
+import { default as _useMap } from 'react-use/esm/useMap.js';
+import { default as _useMountedState } from 'react-use/esm/useMountedState.js';
+import { default as _useNumber } from 'react-use/esm/useNumber.js';
+import { default as _usePrevious } from 'react-use/esm/usePrevious.js';
+import { default as _useQueue } from 'react-use/esm/useQueue.js';
+import { default as _useSet } from 'react-use/esm/useSet.js';
+import { default as _useSetState } from 'react-use/esm/useSetState.js';
+import { default as _useThrottle } from 'react-use/esm/useThrottle.js';
+import { default as _useThrottleFn } from 'react-use/esm/useThrottleFn.js';
+import { default as _useToggle } from 'react-use/esm/useToggle.js';
+import { default as _useUnmount } from 'react-use/esm/useUnmount.js';
+import { default as _useUnmountPromise } from 'react-use/esm/useUnmountPromise.js';
+import { default as _useUpdateEffect } from 'react-use/esm/useUpdateEffect.js';
+import { backgroundOnlyFactory as factory } from './backgroundOnlyFactory.js';
 import {
   useExposureForNode as _useExposureForNode,
   useExposureForPage as _useExposureForPage,
   useStayTime as _useStayTime,
-} from './exposureBased';
+} from './exposureBased/index.js';
 
 export const createMemo = factory(_createMemo);
 export const useBoolean = factory(_useBoolean);